Move help text to metabase-lib
