+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorHelpText.stories.tsx
@@ -3,7 +3,7 @@ import type { ComponentStory } from "@storybook/react";
 import { checkNotNull } from "metabase/core/utils/types";
 import { createMockDatabase } from "metabase-types/api/mocks";
 import { createMockMetadata } from "__support__/metadata";
-import { getHelpText } from "./ExpressionEditorTextfield/helper-text-strings";
+import { getHelpText } from "metabase-lib/expressions/helper-text-strings";
 import ExpressionEditorHelpText, {
   ExpressionEditorHelpTextProps,
 } from "./ExpressionEditorHelpText";
