+++ frontend/src/metabase/query_builder/components/expressions/ExpressionEditorHelpText.tsx
@@ -3,7 +3,7 @@ import { t } from "ttag";
 import TippyPopover from "metabase/components/Popover/TippyPopover";
 import MetabaseSettings from "metabase/lib/settings";
 import { HelpText } from "metabase-lib/expressions/types";
-import { getHelpDocsUrl } from "./ExpressionEditorTextfield/helper-text-strings";
+import { getHelpDocsUrl } from "metabase-lib/expressions/helper-text-strings";
 import {
   ArgumentTitle,
   ArgumentsGrid,
